backend/requirements.txt
Django>=3.2,<4.0
djangorestframework
django-cors-headers
django-environ
django-filter
google-cloud-aiplatform
psycopg
markdown
markdown2

